Change the dither texture's swizzling
BZ: 127120

This is a more elegant way to sample from a float alpha texture.
Instead of sampling from the red channel in the fragment shader
we can set the alpha channel swizzle to redirect it to the
red channel. This lets us sample from the alpha channel in the
fragment shader and get the correct value.

Change-Id: I9ac128c3696b58f3efd3d5edbff2676f240085d4
Reviewed-on: http://android.intel.com:8080/123699
Reviewed-by: Lu, Shenghua <shenghua.lu@intel.com>
Reviewed-by: Zhang, Hongyu <hongyu.zhang@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>