[PORT FROM MAIN] Battery Service: Show lightning on battery icon only on charging
BZ: 112501

Battery icon was showing lightning bolt icon when battery state was
100% full and it was plain when the battery state was 99% full.
There should not be any variation in icon between the two states since
the battery status is the same, ie, 'full'. The aim of this patch is to
show the lightning bolt icon only when the battery-status is 'charging'.

Orig-Change-Id: I8f4ab51cbcd272e64232c0387aab4b68d2ee48f1

Change-Id: I46438c5ba8aeb3db4fc1ce44244c0a02820d1195
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>
Reviewed-on: http://android.intel.com:8080/123794
Reviewed-by: Gopal, Saranya <saranya.gopal@intel.com>
Reviewed-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>