[PORT FROM MAIN] Frameworks: Report right battery status on lock screen.
BZ: 55683

Currently, the UI shows 'charging' in lock screen though the charging
has been stopped due to some reason. But it should actually check
the battery status also to find out if the device is actually 'charging'.
This patch is to show 'plugged in, not charging' if the device is not
charging while plugged in.

Orig-Change-Id: I77f560f9ae5b53f0e88a47beaa4b07ffbad21411

Change-Id: I3a392c5dfc0fa6f6d6eda7f26926c12e0978e3d8
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>
Reviewed-on: http://android.intel.com:8080/123899
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>