[PORT FROM MAIN] Telephony: Fix issue in PhoneNumberUtils.compareLoosely
BZ: 64234

If the number of matched digits is greater than or
equal to MIN_MATCH and if one of the input strings
index has reached the lower limit,
PhoneNumberUtils.compareLoosely should return
true. Currently, it doesn't return true as it is
expecting the index to be < 0 which will not happen
for the following case(country code+number without
"+" prefix) and number without country code)

This patch fixes the issue by checking the index <= 0.

Category: AOSP-Improvement
Domain: Telephony-Base
Severity: Medium
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I3578bbf7ef21508d2550350e85272591ae55e853
Orig-Change-Id: I48e4c79e4949a81902cecf32ff3ffbfadc9c2947
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>