Telephony: Neighboring cell info not returning the correct LAC, CI and PSC
BZ: 129539

With the introduction of the enhanced Neighboring cell
information supporting LTE, some interfaces of NeighboringCellInfo
returns invalid value and also the parceling of the object was faulty.

This patch modifies the public APIs to return the correct
LAC CI and PSC while LTE is supported and also fixes the issue in
parceling the object with LTE support.

Category: AOSP Improvement
Domain: Telephony-Base
Severity: Medium
Origin: Internal
Upstream-Candidate: No, proprietary

Change-Id: I4ed037c0abb69bc675af8db767ecb704dbac6317
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/126688
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: Kraiem, RihabX <rihabx.kraiem@intel.com>
Tested-by: Kraiem, RihabX <rihabx.kraiem@intel.com>