[PORT FROM R42B-STABLE] [widi] Increase RTSP timeout value for HDCP Authentication processing
BZ: 133968

The Android AOSP framework has a timeout of 15 sec for all RTSP negotiation.
This is proving to be insufficient when HDCP authentication is introduced in
the negotiation phase. Miracsat spec states that between sending a RTSP M6 (SETUP)
command and sending a RTSP M7 (PLAY) command, the adapter is entitled to wait
for 9 sec in case HDCP Authentication is going on in parallel.

In this particular case, we observe that Netgear adapter and Samsung adapter
are waiting for 9 sec before sending PLAY command, even though HDCP Authentication
has failed (since DUT does not have production keys yet). Combined with other RTSP
message exchange time, the overall RTSP negotiation time exceeds the 15 sec timeout
set in the framework. Hence the framework times out and tears down the Widi session.

Important thing to note here is that there is no misbehavior here from the adapter's
point of view. It is following the Miracast spec. The problem is that there is no
correlation between the 15 sec timeout set at Android framework level and the actual
RTSP timeouts specified by the Miracast spec. The 15 sec timeout is insufficient when
additional processing (e.g. HDCP) is introduced.

This is not a bug, rather an inter-operability issue and we need to increase the
framework timeout to work properly with all adapters in scenarios like this.

Change-Id: Id27a15cf6d7ac4b28ca00ab5ab26576f762769a7
Orig-Change-Id: I06058698a11fb2a9953ed50396537c737e9093b0
Signed-off-by: yong7X <yongx1.lan@intel.com>
Reviewed-on: http://android.intel.com:8080/128077
Reviewed-by: Veeramani, Karthik <karthik.veeramani@intel.com>
Reviewed-by: Balguri, Mamatha <mamatha.balguri@intel.com>
Reviewed-by: Chowdhury, Rajneesh <rajneesh.chowdhury@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Tested-by: Post, DavidX J <davidx.j.post@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>