[PORT FROM MAIN]EM/THERMAL: Introduce ErrorCorrection temperature
BZ: 126732

In the BYT platform, the skin thermal sensors support interrupt.
To calculate skin temperature from Thermistor temperature,
we use slope/intercept (m & c) values. These are board specific.
skin temperature (y) = mX + C (where X is the thermistor temperature)
The thresholds that are provided in the thermal_sensor_config.xml files
are 'skin' thresholds and not 'thermistor' thresholds.
When the user space programs these thresholds, the driver converts
them to appropriate 'thermistor' thresholds (using x = (y-c)/m)
and then programs them into the ADC. Hence, the thresholds configured
in the XML are not _exactly_ same as those that get programmed
into the ADC registers. There is some amount of error introduced
due to this conversion. The driver on receiving the interrupt sends
a UEvent to user space with name, temperature, event type etc as
parameters. The Thermal service extracts the current temperature
from the UEvent and compares it with the thresholds configured
in the XML. Depending on the result of the comparison,
(either the temperature went above/below some 'state')
throttling actions are taken.

Sometimes, (due to above mentioned) conversion errors, the current
temperature may still be lower than the threshold configured
in the XML, when an interrupt is generated. In this case,
when the user space compares the threshold, it _may not_ see the
threshold being _crossed_ in higher direction; and miss the 'event'.
This will result in a throttling action being missed.

To get this issue fixed, we introduce an 'ErrorCorrectionTemperature'
which is the maximum difference between the threshold in XML and
the threshold programmed in the ADC. When the user space receives a
UEvent, it adds this offset to the temperature, so that no events are
missed.

Orig-Change-Id: I60db6260dcadde708ed19841aceb3253dbdf040e

Change-Id: Ib272a6f32d4973996868b57f6fa661d625fe3735
Signed-off-by: rjayapax <ranjinix.jayapal@intel.com>
Reviewed-on: http://android.intel.com:8080/124955
Reviewed-by: Roychowdhury, Saikat <saikat.roychowdhury@intel.com>
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>