Integrate Arkham to mainline
BZ: 127460

Arkham is a deeply integrated security container capability for Android for the
purpose of isolation and management in mixed used cases. Primarily Arkham is
intended to enable separation of work and personal content and capability.

Detailed information about the Arkham requirements and architecture can
be found here: https://sharepoint.amr.ith.intel.com/sites/SDI/Arkham/

This patch adds framework modifications to implement Arkham functionality like
the ones highlighted below. Most of the modifications are done in a Parent or
Extend class and this patch only contains the modifications to hook into the
framework classes. The full implementation can be found in
vendor/intel/arkham/frameworks/enabled.

() Containers - are a special class of users which provides IPC and FS isolation
as well as FS encryption via ecryptfs. One important difference between
containers and Android users is that container applications run at the same time
as the primary user's applications, no switch is required to run container
applications.

() Container IPC isolation is provided by the default Android users
mechanism. Several tweaks are made to allow IPC communications for
some use cases such as resolving Intents from inside container
applications to outside container applications (but not the reverse),
policy based copy-paste, etc.

() Filesystem isolation is provided by setting the container data
directory rigths to 0770 and users.group to system.ContainerGID and
running all container applications with the ContainerGID in its
supplementary groups. See more in the patch against system/core.

() Container Launcher - it is the application that present the interface to the
user with the container. Its looks and feel is similar with the native launcher
and its purpose is to present to the user the applications installed in the
container and to allow the user to launch them.

() Identification of container applications - container applications are
identified in the recent applications list with the container name listed under
the application name as well as with the container icon super imposed on the
application icon.

() Disable screenshots for container applications

() Controlled copy-paste between container applications and primary user
applications. There are policies that the container administrator can set to
allow inbound or outbound copy-paste operation to/from the container.

() Contacts merging. Three contact merging policies are supported: no merging,
"normal" and "encrypted". "Normal" merging allows the primary user to see some
fields of the container contacts. These contacts are not encrypted and stored
outside of the container but they can be accessed with unlocking the
container. The "encrypted" policy allows the primary user to see container
contacts, they are stored in the container and thus are encrypted, but the user
needs to unlock the container before accessing them.

() Keyguard for container and container applications. The container
administrator can set a different authentication policy (pin, password, pattern)
then the device authentication. In order to access the Container Launcher the
user needs to unlock the container by entering the authentication
pin/pattern/password. Also, when switching to a container application, if the
container is locked the user needs to re-authenticate. The container locks
itself after a period of inactivity set by the administrator.

() Package Manager changes to allow skiping touching /data directories
for containers until the user unlocks the container and thus the
container underlying FS is decrypted. Also new FileObserver added for
the /data/containers to enable new Container Launchers as new
containers are created.

Category: feature differentiation
Domain: AOSP.Extend-Arkham
Origin: Internal
Upstream-Candidate: No, proprietary

Change-Id: I931c3377781c280a24d72797de5e7d297f18a314
Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>
Reviewed-on: http://android.intel.com:8080/126919
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>