Telephony: Fix the issue that "+86 " can't be removed
BZ: 124157

When input a "+86 " in chinese, it will be formatted to "+86 ". Then deleting
then last character " ", it will cause that the result "+86" is auto formatted
to "+86 ". It will show "+86" can not be deleted.

Here deleting/adding a character will check if the phone number can be re-formatted
to original one. If true, stop formatting.

Category: AOSP Improvement
Domain: Telephony
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I7f563f484099c9b27fb5158cffe22282550e862b
Signed-off-by: Zhang Dongxing <dongxing.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/126603
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Liu, JianX <jianx.liu@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>