WiFi Display: delete p2p group when forgetting display
BZ: 98216

When a wifidisplay is forgotten in WifiDisplay Settings activity,
the coresponding p2p group is not removed.

In order to delete the p2p group associated to the wifidisplay, the
function that updates the list of group in WifiP2pService has been
modified so that it also updates the client list for each group.
Thus when a display is forgotten, WifiDisplayController retrieves
the group list from  WifiP2pService, finds the group in which the
display was associated and deletes it.

Category: aosp improvement
Domain: CWS.WIDI-Framework
Origin: internal
Upstream-Candidate: yes

Change-Id: Ibc4066d9686edd297b89d5d03bc9e855ac53d96c
Signed-off-by: Cedric IZOARD <cedricx.izoard@intel.com>
Reviewed-on: http://android.intel.com:8080/127340
Reviewed-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>