EM/THERMAL: Notify user of display brightness throttle
BZ: 133533

This patch enables thermal manager to notify user of a
display throttle event due to higher thermal states of
skin. User is notifed via a vibrate and a toast message.
Context is passed as input to all the cooling device init
metod, so that any device can use the ThermalNotifier class
in future to generate user notifications.

Change-Id: I8932a2a799a071d9156f38aeca15e6d7e85b3641
Origin-Change-Id: I3b85c8bfc6cacb32194f7e5ad1677fd507c007cf
Signed-off-by: Saikat <saikat.roychowdhury@intel.com>
Reviewed-on: http://android.intel.com:8080/134445
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Kundella, SreedharX <sreedharx.kundella@intel.com>
Reviewed-by: Jayapal, RanjiniX <ranjinix.jayapal@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>