[PORT FROM MAIN JB-MR2] Wifi: Adjustments for sleep policy
BZ: 128884 26274 15926 17650 28366

The wifi sleep policy "When screen turns off" will switch off wifi although the USB tethering is active.
With this patch, the sleep policy is removed when the usb tethering is active.
The WifiService maintains a local status variable in according with the wifi tethering (on,off).
If the usb tethering is on, the alarm, to switch off the wifi, is cancelled.

If the device in unplugged after turning off the
display while still not connected, it's better to directly turn off wifi in
order to reduce power consumption.
So when unplugged while screen is already off, request to go in idle (and so
turns off wifi) if wifi is not connected.
When USB tethering is enabled, don't care about sleep policy:
  - first because DUT is powered by USB
  - secondly because DUT is used as wifi card and so used by PC

Change-Id: Ic1af2367a2f73965c32d300450ee76cf5593af78
Orig-Change-Id: I2adc1228ffa6d1b73663966656c2297201d001f6
Category: aosp improvement
Domain: CWS.WIFI-Sleep_Policy
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Paul Naoumenko <paulx.naoumenko@intel.com>
Reviewed-on: http://android.intel.com:8080/131437
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Tested-by: cactus <cactus@intel.com>