[GNSS] Don't send CELLID when END msg has been sent
BZ: 126349

In some cases, requestLocationUpdate() can take several seconds to report
the CELLID informations. If the information takes more than 3 seconds to arrive,
a timeout is triggered and the END message is sent. In this case, we must ensure
that the CELLID information is not sent to the native stack.

Timeout is lowered to 3 seconds to improve the GPS performance when in this
particular case.

Category: aosp improvement
Domain: CWS.GPS-MultipleLocationId
Origin: internal
Upstream-Candidate: yes

Change-Id: I9665ba861dd35ddfd503b967bea4b5091bccb13a
Signed-off-by: Guillaume Aubertin <guillaume.aubertin@intel.com>
Reviewed-on: http://android.intel.com:8080/132924
Reviewed-by: Peix, Fabien <fabien.peix@intel.com>
Reviewed-by: Suet, Nicolas <nicolas.suet@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Peng, QingwenX <qingwenx.peng@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>