[PORT FROM MAIN][GPS] Don't send CELLID when END msg has been sent
BZ: 126349

In some cases, requestLocationUpdate() can take several
seconds to report the CELLID informations.
If the information takes more than 3 seconds to arrive,
a timeout is triggered and the END message is sent.
In this case, we must ensure that the CELLID information
is not sent to the native stack.

Timeout is lowered to 3 seconds to improve the GPS
performance when in this particular case.

Change-Id: I8fb97eeccaffdc9bbf5bc65101ac5c1e954160cb
Orig-Change-Id: I9665ba861dd35ddfd503b967bea4b5091bccb13a
Category: feature differentiation
Domain: CWS.GPS-SUPL
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Guillaume Aubertin <guillaume.aubertin@intel.com>
Signed-off-by: jerome Pantaloni <jeromex.pantaloni@intel.com>