[PORT FROM MAIN]wifi: Do not remove certificate when network is forget [WORKAROUND]
BZ: 134286

Remove the certificate removal from AP forget process otherwise
the certificate is no more available for a further connection.

Tagged as WORKAROUND because it's not clear if this a expected behavior
or an issue due to Android code change. This change was introduced
on Android 4.3 with eap-enterprise code re-organization.

Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: no,proprietary

Change-Id: Ibbed7e659583ac279674c3a888a38e2f602dd5bc
Orig-Change-Id: Ie203c1a871ebfe6070a14fe4339760ac6cbab694
Signed-off-by: Honore Tricot <honorex.tricot@intel.com>
Reviewed-on: http://android.intel.com:8080/134380
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Reviewed-by: Izoard, CedricX <cedricx.izoard@intel.com>
Reviewed-by: Regairaz, Julie <julie.regairaz@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>