Add HeartBeatHandler debug info for debug purpose
BZ: 140949

This change allows to analyze the UIWDTs starting from the right thread.
This is because MR1 introduced a change in the thread handling the
HearBeatHandler; it was originally com.android.ServerThread, then it
changed to ActivityManager and now, in MR2, it might be ActivityManager or
android.server.ServerThread.

This change will allow us to put in system_server_watchdog@xxx the right
information systematically; even it that changes again.

Category: feature differentiation
Domain: Services-Core

Origin: internal
Upstream-Candidate: no, proprietary

Change-Id: Ie700feb358d760496efb09c2f146208c0d28f0d3
Signed-off-by: Jean-Christophe PINCE <jean-christophe.pince@intel.com>
Reviewed-on: http://android.intel.com:8080/134497
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>