EM/THEMAL: Modify the init function in GPUMaxFreqControl
BZ: 144356

[Description of the issue]
The GPU cooling device was not getting throttled during
a temperature zone's state change.

[Implemented Solution]
The init function in the GPUMaxFreqControl file
had to be changed since a new argument was added
in the calling side. This was introduce by an
earlier patch and this got missed.

Category: feature differentiation
Domain: EM.ThermalManagement-commmon
Origin: internal
Upstream-Candidate: no, not AOSP

Change-Id: I632defa0d8b6722c64a4069f60aa8b1eda2aa4ac
Signed-off-by: Ajay <ajay.thomas.david.rajamanickam@intel.com>
Reviewed-on: http://android.intel.com:8080/137779
Reviewed-by: Pawnikar, Sumeet R <sumeet.r.pawnikar@intel.com>
Reviewed-by: B M, Shravan K <shravan.k.b.m@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Yin, Fengwei <fengwei.yin@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>