Prevent user from doing shutdown during phone encryption
BZ: 146550

When the user perform a graceful shutdown during encryption,
the shutdown takes around 15 minutes to finish.

Since the shutdown during the encryption corrupts the phone
and force the user to perform a factory reset, it seems better
not to allow the user to make a shutdown during the encryption.
He can still perform an hardware shutdown and the result is
the same (need to reset the phone).

Category: aosp improvement
Domain: PUPDR

Origin: internal
Upstream-Candidate: yes

Change-Id: I42346f2fab6ba09778977b613b70b59dd5427bc8
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>