EM/PowerUI: Reduce frequency of charger notification
BZ: 128877

Problem: The user is notified when the battery level drops
when the charger is connected. It could be annoying to the
user since the toast message comes for every percentage drop
in battery level as long as the charger is connected.

Solution: "Insufficient charger current" toast should be
displayed only once as long as the charger is connected.
The message will be displayed again only when there is
a change again in battery status (like discharging or
not charging) and again status comes back to 'charging'
still consuming more current leading to battery level
drop.

Category: feature differentiation
Domain: EM.Battery
Origin: internal
Upstream-candidate: no, proprietary

Change-Id: I11f0252cabe2cbdddef197ed4563e0c98c0da227
Signed-off-by: Saranya Gopal <saranya.gopal@intel.com>