Arkham bugfixes WW43
BZ: 136675

* Fix disabling container when maximum nr of failed attempts is reached
* Re-implement screenshot restriction, using secure layers and removing the old code
* Container contacts not listed in Primary user when contacts merging is set to NORMAL
* Fix displaying WiFi networks in container Settings
Added the check for container users inside WifiService, to allow
displaying WiFi networks in the container Settings app.

* Implement copy / paste functionality in Extend ClipBoard Service
This implementation also calls super to maintain default behavior.

* Fix container icon overlay for normal contacts
Removed the ParentQuickContactBadge's assignContactUri method, as it
wasn't necessary to determine if the contact belongs to a container
or not. This is determined by the onContactUriChanged method.

* Replace the Back key with the Home key
The Saltbay Back key replaced with the Home key while KeyGuard is activated in the Container.

* Fix SystemUI crash caused by Gmail unparcelable extra

* Fix uiautomator in container
If the callingUid is the root, resolve it by using the current user set in accessibility service.

* Add Arkham comment to keyguard change
Replaced a hardcoded number with a system constant to improve code readability

* Set keyguard softkeys based on user type
The keyguard will now refresh its system softkeys based on
what type of user the keyguard belongs to. Example: hide all
buttons for the primary user, but show the home button for
the container.

* Fix IntelAppsLauncher present in container
Intel Apps Launcher is present inside the container just after the
container was created. After a while it dissapears. While it is present,
it is useless.

* Contacts not getting deleted after container is wiped or disabled
* Do not silently reboot the phone when unmount fails
When removing or disabling a container, the device is rebooted
if unmounting ecryptfs fails.
Instead of immediately rebooting the phone, send an intent to
MDM. MDM can ask for user's permission and then reboot the phone.
Normal applications do not have rights to reboot the phone, so
a call to ContainerManagerService is needed for that. ContainerLauncher
is a system application and could reboot the phone, but it is deleted
when the container is launched (so we cannot send an intent to
ContainerLauncher is unmount fails when removing the container).

Category: feature differentiation
Domain: AOSP.Extend-Arkham
Origin: Internal
Upstream-Candidate: No, proprietary

Change-Id: I43483ebbc3afd20c702600b971fb9a955e3d2a36
Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>
Signed-off-by: Irina Tirdea <irina.tirdea@intel.com>
Signed-off-by: Arvind Mohan <arvind.mohan@intel.com>
Signed-off-by: Rares Petrescu <rares.petrescu@intel.com>
Signed-off-by: Viorel Suman <viorel.suman@intel.com>
Signed-off-by: Tiberiu Breana <tiberiu.a.breana@intel.com>
Signed-off-by: Aman Bahshi <aman.bakshi@intel.com>
Signed-off-by: Adriana Reus <adriana.reus@intel.com>
Signed-off-by: Mihai Serban <mihai.serban@intel.com>