ASF 1.0: Resolve Klocwork issue in launchAsf.
BZ: 146969

Klockwork found the issue "Dereference of a null value which is
returned from a method".  This was resolved by checking to make sure
the intent was not null before dereferencing.

Since this issue was reported on a platform that has ASF disabled, the
logic of launchAsf() was changed so the entire body is within 'if
(AsfAosp.ENABLE)'.

Category: feature differentiation
Domain: AOSP.Security-ASF
Origin: internal
Upstream-Candidate: no: proprietary
Change-Id: Idecbdefff922db51bb3a8c80b944a584b57089ff
Signed-off-by: Norvin Shuster <norvin.shuster@intel.com>