EM/THERMAL: Fix Intent info sent out by Modem Zone
BZ: 143912

[Description]
In thermal intent, the state sent is that max sensor state,
but the temperature sent is not corresponding to the sensor
with max state.

[Implemented Solution]
Send the Sensor with max state and the temperature of that
particular sensor in Thermal Intent.

Category: feature differentiation
Domain: EM.ThermalManagement-commmon
Origin: internal
Upstream-Candidate: No, Not AOSP

Change-Id: I5934818563b4046d45e36e22e8d724c42c5dba9f
Signed-off-by: Saikat <saikat.roychowdhury@intel.com>