[PORT FROM MAIN] Shutdown: In airplane mode, shutdown modem using oem hooks
BZ: 123894

In airplane mode, radio is off but the modem can
still be powered on due to CWS clients accessing
the modem. If the platform shutdown is initiated
in this case, modem is not powered off properly.

On platform shutdown in airplane mode(radio off),
use oem hooks to power off the modem.

Category: AOSP Improvement
Domain: Telephony-Base
Origin: Internal
Upstream-Candidate: No, Proprietary

Change-Id: Id699833bd4cd6e08b90a9132383df85bf9f11787
Orig-Change-Id: Id177c39f2ba881176f9ee600f849514a9f6cbcc6
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>