[PORT FROM MAIN] BatteryStatsImpl: Avoid memory corruption due to Parcel Recycling
BZ: 47466

While processing batterystatsfile, in present code there is a race where a parcel
might be recycled while it is being written to disk in another thread.
The write process is done in 2 steps:
- write into a Parcel (memory block)
- then write this Parcel to disk using Journaled API.
The Journaled API ensures robustness against file-system issues due to power cut.
The writing is done in batch using a thread running commitPendingDataToDisk().
This function writes the parcel to disk but this parcel is a global variable
initialized by writeLocked(). If the writing takes too long time and two write
threads are launched, the parcel is freed by the second writeLocked() call;
which could cause the memory corruption.

Change global Parcel variable by a Parcel parameter to commitPendingDataToDisk()
and only this function will be allowed to recycle the parcel.

Category: device enablement
Domain: EM.BatteryManagement-Common
Origin: internal
Upstream-Candidate: yes

Orig-Change-Id: I1884141d7d9c61601e2c8e07523a6e9aa6709afc

Change-Id: I9f7895d19ee64d14213bd2c15fb3d6bc4fad8575
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>