[PORT FROM MAIN] Frameworks: Report right battery status on lock screen.
BZ: 55683

Currently, the UI shows 'charging' in lock screen though the charging
has been stopped due to some reason.

Battery status should also be checked to find out if the device is
actually 'charging'. This patch is to show 'plugged in, not charging',
if the device is not charging while plugged in.

Category: device enablement
Domain: EM.BatteryManagement-Common
Origin: internal
Upstream-Candidate: no, proprietary

Orig-Change-Id: I3a392c5dfc0fa6f6d6eda7f26926c12e0978e3d8

Change-Id: I9d4eb92a09491f4f1675a5ce7829a4b2c12c9173
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>