[PORT FROM MAIN JB-MR2] Accelerate the thumbnail extraction when the sync-frame duration is long
BZ: 140942

When the sync frame duration is very long in the test clips, the
thumbnail extration will be very slow. Because the program will
have to decode from the previous sync-frame to the desired frame.
If end user presses the exit button before the thumbnail
extraction finishes, sometimes ANR will happen.

In this patch, if the desized frame's timestamp is 1secs larger
than the previous sync-frame, we will choose to use the sync-frame
as the thumbnail.

Category: aosp improvement
Domain: Video.Media-video editor
Origin: internal
Upstream-Candidate: yes
Change-Id: Ia25a1e688cf1583ee4cf614518d5c5e5d304c3d9
Orig-Change-Id: I0914ea45bb4abf34da65ac045e2a4688e3b8f4d4
Signed-off-by: wfeng6 <wei.feng@intel.com>
Signed-off-by: Dan Liang <dan.liang@intel.com>