[PORT FROM MAIN] PowerUI: Notify user during discharge while charger is connected
BZ: 106140

Sometimes, the current supplied by the charger will be lesser than the current
consumed by the platform. User should be notified in this case.

This patch notifies the user when the capacity drops while the charger is plugged in.
It is displayed only once as long as the charger is connected.

Orig-Change-Id: I6b03529e52d412f09af1b75495f669928e702530
Category: device enablement
Domain: EM.BatteryManagement-Common
Origin: internal
Upstream-Candidate: no, proprietary
Orig-Change-Id: I492c90b949326c7c66dd18371ab9fb74769f84c0
Orig-Change-Id: I3547e835d9f8b5c860112aa6efeb516a17036278
Orig-Change-Id: I11f0252cabe2cbdddef197ed4563e0c98c0da227

Change-Id: If63bb790a35f9d6bc3424484263307ab38649c87
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>