[PORT FROM MAIN] PowerUI: Turn-on display at low battery warning
BZ: 52107

When Screen is off, then low battery popup is not visible to user.

This patch turns on display just before popup and it will remain
in this state for 3 sec.
Also, if in silent-mode, then vibra will be turned on for 0.5 sec.

Category: device enablement
Domain: EM.BatteryManagement-Common
Origin: internal
Upstream-Candidate: yes

Orig-Change-Id: I618c8544229adb71870010f26407022450137a98

Change-Id: I0b1cd10a0caa9bdf5e53f082d31452d9c2e5ad7c
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>