Fix the LONG_PRESS_POWER_NOTHING behavior
BZ: 147702

When the LONG_PRESS_POWER_NOTHING option is selected in
mPowerLongPress function, the variable mPowerKeyHandled is not set
to True. This leads to an error when pressing on power button for
2s because the mPowerLongLongPress callback is then never cancelled.
So there is a graceful shutdown detected after 6s even if power
button has been released.

This patch factorize the code so the mPowerKeyHandled variable is
set to True at the very beginning of the function to ensure that
pending callback will be correctly cancelled whatever the chosen
behavior.

Category: feature differentiation
Domain: PUPDR.Subsystem2-Framework-core
Origin: internal
Upstream-Candidate: no, not-aosp
Change-Id: I7c5150c8f296a380edde8a2e6bb2396fe6886f61
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>