[PORT FROM MAIN] BatteryService: Show lightning on battery icon only on charging
BZ: 112501

If Battery icon shows lightning symbol even after the device has charged fully
or when the battery is actually not charging though charger is plugged in,
the user might feel that the device is still charging.

Remove the lightning icon as soon as the battery reaches 'FULL' or if it is
not charging while charger is plugged in.
Also, there should not be any variation in icon between the two states where
either 100% full or less than 100% full.

Change-Id: I6d3a2e47fc159e3e6fb493298627b4304809b04e
Orig-Change-Id: I46438c5ba8aeb3db4fc1ce44244c0a02820d1195
Orig-Change-Id: Ie65946a6f0502cd309f91ba25ecc3b1ddc0aef20
Category: device enablement
Domain: BatteryManagement-Common
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>
Signed-off-by: Saranya Gopal <saranya.gopal@intel.com>