Integrate Arkham to mainline
BZ: 156819

Arkham is a deeply integrated security container capability for Android for the
purpose of isolation and management in mixed used cases. Primarily Arkham is
intended to enable separation of work and personal content and capability.

Detailed information about the Arkham requirements and architecture can
be found here: https://sharepoint.amr.ith.intel.com/sites/SDI/Arkham/

This patch adds framework modifications to implement Arkham functionality like
the ones highlighted below. Most of the modifications are done in a Parent or
Extend class and this patch only contains the modifications to hook into the
framework classes. The full implementation can be found in
vendor/intel/arkham/frameworks/enabled.

* Containers - are a special class of users which provides IPC and FS isolation
as well as FS encryption via ecryptfs. One important difference between
containers and Android users is that container applications run at the same time
as the primary user's applications, no switch is required to run container
applications.

* Container IPC isolation is provided by the default Android users
mechanism. Several tweaks are made to allow IPC communications for
some use cases such as resolving Intents from inside container
applications to outside container applications (but not the reverse),
policy based copy-paste, etc.

* Filesystem isolation is provided by setting the container data
directory rigths to 0770 and users.group to system.ContainerGID and
running all container applications with the ContainerGID in its
supplementary groups. See more in the patch against system/core.

* Container Launcher - it is the application that present the interface to the
user with the container. Its looks and feel is similar with the native launcher
and its purpose is to present to the user the applications installed in the
container and to allow the user to launch them.

* Identification of container applications - container applications are
identified in the recent applications list with the container name listed under
the application name as well as with the container icon super imposed on the
application icon.

* Disable screenshots for container applications

* Controlled copy-paste between container applications and primary user
applications. There are policies that the container administrator can set to
allow inbound or outbound copy-paste operation to/from the container.

* Contacts merging. Three contact merging policies are supported: no merging,
"normal" and "encrypted". "Normal" merging allows the primary user to see some
fields of the container contacts. These contacts are not encrypted and stored
outside of the container but they can be accessed with unlocking the
container. The "encrypted" policy allows the primary user to see container
contacts, they are stored in the container and thus are encrypted, but the user
needs to unlock the container before accessing them.

* Keyguard for container and container applications. The container
administrator can set a different authentication policy (pin, password, pattern)
then the device authentication. In order to access the Container Launcher the
user needs to unlock the container by entering the authentication
pin/pattern/password. Also, when switching to a container application, if the
container is locked the user needs to re-authenticate. The container locks
itself after a period of inactivity set by the administrator.

* Package Manager changes to allow skiping touching /data directories
for containers until the user unlocks the container and thus the
container underlying FS is decrypted. Also new FileObserver added for
the /data/containers to enable new Container Launchers as new
containers are created.

* Fix JIRA: ARKHAM-1268
The reported klockwork issues:
https://android-kw2.jf.intel.com:8090/review/insight-review.html#goto:project=mainline_redhookbay,pid=107404
https://android-kw2.jf.intel.com:8090/review/insight-review.html#goto:project=mainline_redhookbay,pid=107411

* Fix disabling container when maximum nr of failed attempts is reached
* Re-implement screenshot restriction, using secure layers and removing the old code
* Container contacts not listed in Primary user when contacts merging is set to NORMAL
* Fix displaying WiFi networks in container Settings
Added the check for container users inside WifiService, to allow
displaying WiFi networks in the container Settings app.

* Implement copy / paste functionality in Extend ClipBoard Service
This implementation also calls super to maintain default behavior.

* Fix container icon overlay for normal contacts
Removed the ParentQuickContactBadge's assignContactUri method, as it
wasn't necessary to determine if the contact belongs to a container
or not. This is determined by the onContactUriChanged method.

* Replace the Back key with the Home key
The Saltbay Back key replaced with the Home key while KeyGuard is activated in the Container.

* Fix SystemUI crash caused by Gmail unparcelable extra

* Fix uiautomator in container
If the callingUid is the root, resolve it by using the current user set in accessibility service.

* Add Arkham comment to keyguard change
Replaced a hardcoded number with a system constant to improve code readability

* Set keyguard softkeys based on user type
The keyguard will now refresh its system softkeys based on
what type of user the keyguard belongs to. Example: hide all
buttons for the primary user, but show the home button for
the container.

* Fix IntelAppsLauncher present in container
Intel Apps Launcher is present inside the container just after the
container was created. After a while it dissapears. While it is present,
it is useless.

* Contacts not getting deleted after container is wiped or disabled
* Do not silently reboot the phone when unmount fails
When removing or disabling a container, the device is rebooted
if unmounting ecryptfs fails.
Instead of immediately rebooting the phone, send an intent to
MDM. MDM can ask for user's permission and then reboot the phone.
Normal applications do not have rights to reboot the phone, so
a call to ContainerManagerService is needed for that. ContainerLauncher
is a system application and could reboot the phone, but it is deleted
when the container is launched (so we cannot send an intent to
ContainerLauncher is unmount fails when removing the container).

Change-Id: I0ba79cb60be0776c4ab4b1224aefd6779680bdad
Category: feature differentiation
Domain: AOSP.Security-Arkham
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>
Reviewed-on: http://android.intel.com:8080/126919
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: cactus <cactus@intel.com>