Arkham bugfixes WW47
BZ: 156819

* Delete Container APKs while wiping container.
apk is not deleted after wiping container if it is installed only inside the container.

* Fix Keyguard password counter
Fixed a bug where the Keyguard was displaying an erroneous
number of failed password attempts, giving the impression
that it disabled the container after only two failed unlock
attempts.

* Fix password expiration notification bug
Changed the DevicePolicyManager's handlePasswordExpirationNotification
method access to protected, so that we can override it in the Extended DPM.
This method is called after the container is unlocked and the BOOT_COMPLETED
intent is sent. In the EDPM, we will reload the container device admin
info before handling password expiration notifications.

* Disable keyguard if FLAG_KEEP_SCREEN_ON is set
FLAG_KEEP_SCREEN_ON is held by activities which don't want to turn the screen off.
Examples are Youtube, Video player etc.
Disable keyguard if this flag is set so that videos can be played without keyguard interruption.

* Fix NPE reported by Klockwork
The reported Klockwork issues:
https://android-kw2.jf.intel.com:8090/review/insight-review.html#goto:project=mainline_merr_vv,pid=122019
https://android-kw2.jf.intel.com:8090/review/insight-review.html#goto:project=mainline_merr_vv,pid=122020
https://android-kw2.jf.intel.com:8090/review/insight-review.html#goto:project=mainline_merr_vv,pid=122021

* Allow Bluetooth management and sharing in container
Allow container to process intents sent from Bluetooth service.

* Set container userid for container notification resources
Currently, notifications coming from container applications have
their userId changed to their owner's id. In case we have an
application installed only in a container, the StatusBar code would
try and load application resources from the primary user instead of
the container user, causing exceptions to be thrown.
This change replaces some Notification userid fields in SystemUI code,
where application resource fetching is required, with the container
userid, if this id has been written in the Notification's 'tag' field.
The tag field is written by the NotificationManagerService when it
changes a container notification's user handle from container to primary.

* Camera policy hotfix
Fixed a bug where setting the camera_disabled policy to 'true' for a
container would cause the camera to be disabled (crash) for the primary
user as well.

Change-Id: Iba73f74da71bb8ff3f99044bd034d7ac4e383a2f
Category: feature differentiation
Domain: AOSP.Security-Arkham
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Tiberiu Breana <tiberiu.a.breana@intel.com>
Signed-off-by: Arvind Mohan <arvind.mohan@intel.com>
Signed-off-by: Aman Bakshi <aman.bakshi@intel.com>
Signed-off-by: Rares Petrescu <rares.petrescu@intel.com>
Signed-off-by: Mihai Serban <mihai.serban@intel.com>