Revert "[PORT FROM MAIN] widi: Allow background video player with mipi off"
BZ: 158343

This reverts commit 4658ccb7dd2c7c88c42ad892a2a320d64682d6bf.

This ported patch didn't do what it was supposed to do, due to some
porting errors and also a new change in Android 4.4.1 that wasn't
accounted for in the original patch. I'm reverting it in order to have
one clean patch in the history for this feature.

Change-Id: Ic3184d42612e5d092509eb39fd21b7dcd9301c53
Category: feature differentiation
Domain: Video.Rendering-WIDI
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Brian Rogers <brian.e.rogers@intel.com>