EM/THERMAL: fix for shutdown at the end of emergency call
BZ: 119255

[Description]
When emergency call is ongoing, if a zone is in CRITICAL state,
platform is not allowed to shutdown. However, if the zone comes
back to thermal state less than CRTICAL, platform should not shut
down when emergency call ends.

[Implementation]
A map is maintained for each zone which stores a TRUE/FALSE
value indicating if there is a pending critical shutdown.
A count variable is maintained indicating total number of zones
with map value as true. At the end of an emergecny call, if the
count is > 0 platform shuts down else no action is taken.

Change-Id: Iad6bac31888aad046acff2b9fa9a12b0bffa6459
Category: device enablement
Domain: ThermalManagement-Common
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Saikat <saikat.roychowdhury@intel.com>