Wifi: Start CwsServiceManager before CsmWifiOffloadSystemService
BZ: 158483

CwsServiceManager is now a system service but it is started after
CsmWifiOffloadService and when CsmWifiOffloadSystemService starts
it creates a CsmClient. During its creation the CsmClient cannot
binds to CwsServiceManager as it is not yet started, and then
this CsmClient cannot be used by CsmWifiOffloadSystemService.

This patch starts CwsServiceManager first.

Change-Id: Ib1ace9fc5d31431a73f1f69cd55f79ea677e2ef5
Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Cedric IZOARD <cedricx.izoard@intel.com>