widi: Allow background video player with local display off
DO NOT UPSTREAM

The new background video player uses the standard Android Presentation
API to send its content to the remote display. A Presentation is
normally created using an Activity context, and it gets closed when
the activity is no longer in the foreground. To run a Presentation in
the background, we instead start the Presentation from a Service
provided by the player app.

Normally creating windows from the service context would be blocked,
but if you get the SYSTEM_ALERT_WINDOW permission and set a Dialog's
type to TYPE_SYSTEM_ALERT, you can create it. Since Presentation is a
subclass of Dialog, we can do the same thing with Presentations.

But here we are introducing new type of window similar to TYPE_SYSTEM_ALERT
as TYPE_INTEL_SYSTEM_ALERT. Also changed policy to map this window to
SYSTEM_ALERT_WINDOW permission. And the priority of this window type
is same as TYPE_SYSTEM_ALERT.

However, if you turn the phone's display off with the power button,
you'll see a black screen on the remote display instead of the app's
content. There are three reasons for this:

1. There is logic in the window manager that forces external displays
   into 'mirror' mode when there is a window of type TYPE_DREAM or
   TYPE_KEYGUARD present, even if there is unique content intended for
   the external display. The lock screen (TYPE_KEYGUARD) is considered
   present while the screen is off, so it puts the widi display into
   mirror mode.

2. There is global state tracking whether all displays are on or all
   displays are off. If all displays are considered to be off, the
   display manager sets an isBlanked parameter to true in a call to a
   LogicalDisplay method, which makes it set its layer stack id to
   BLANK_LAYER_STACK. This means no layers are sent to SurfaceFlinger
   for this display.

3. As of Android 4.4.1, there is now a Presentation created by the
   lock screen to protect privacy during PIN entry. Its type is
   TYPE_KEYGUARD_DIALOG, which normally stacks above our Presentation
   with TYPE_SYSTEM_ALERT. This is present while the screen is off and
   during unlocking, covering up our background Presentation.

For the first two problems, I have WindowManagerService set a flag
when it sees a window of TYPE_INTEL_SYSTEM_ALERT on something other
than the primary display. When the flag is set, the associated display
will be considered to have unique content even when there is a keyguard
present. Then, when the display manager sees that a background
Presentation is present, it will set isBlanked to false even if the
global blanked state is true, ensuring that the Presentation remains
visible even when the local display is off.

For the third problem, I have given adjusted background presentations
(with window type TYPE_INTEL_SYSTEM_ALERT) that are on external display
so that they are one layer above the normal TYPE_KEYGUARD_DIALOG windows.

Change-Id: I6f851852ec339d5107174e8f15e80e61dfc6401b
Category: feature differentiation
Domain: Video.Rendering-WIDI
Origin: internal
Upstream-Candidate: no, proprietary
Issue: AIAF-5318
Signed-off-by: Sanjay Aghara <sanjay.r.aghara@intel.com>
Reviewed-on: https://android.intel.com/160199
Reviewed-by: Suresh, Anitha <anitha.suresh@intel.com>
Tested-by: Suresh, Anitha <anitha.suresh@intel.com>
Reviewed-by: Fagerstedt, Axel <axel.fagerstedt@intel.com>
Reviewed-by: Chiras, Robert <robert.chiras@intel.com>
Tested-by: sys_abtbuild <sys_abtbuild@intel.com>