Fix Watchdog HeartbeatHandler to run on correct thread
Issue: BMDA-393

The HeartbeatHandler for the System Server Watchdog has been running
on the wrong thread due to a race condition in initialization.  It's
designed to run on ServerThread, so that it can catch lockups in the
main looper of the System Server.  It has been running on
ActivityManagerThread instead, so it does not detect lockups on the
ServerThread as it should.

ActivityManagerService is calling Watchdog.getInstance() before
ServerThread calls Watchdog.getInstance().init(), so the handler is
being bound to the ActivityManagerThread instead of the ServerThread.

Explicitly bind HeartbeatHandler to ServerThread, so that the Watchdog
catches lockups on this critical thread.

cherry-pick from https://android-review.googlesource.com/#/c/54042

Author	John Michelau <w14107@motorola.com>
Signed-off-by: Yong Zhi <yong.zhi@intel.com>

Change-Id: I9be9df6d4f8eee22304ddb75581fe36226df3b9a