Disable soft keyboard for the PIN input box
Two number keyboards are displayed when tapping the PIN input box
in the lock screen if the "Password" option was selected and the device
was locked and unlocked at least for 1 time before having the "PIN"
option set for the Screen lock.
Fix this issue by hiding the soft keyboard when the user touches
the PIN EditText.

Issue: APDEV-1073
Category: AOSP Improvement
Domain: AOSP-apps
Origin: internal
Upstream-Candidate: Yes
Change-Id: I53468e2caef460831dcdc50d8c196565719aa9d5
Signed-off-by: Constantin Musca <constantin.musca@intel.com>